refactor: move estimation services into services directory
