+++ src/app/components/assumptions-and-limitation/assumptions-and-limitation.component.ts
@@ -3,9 +3,9 @@ import { DecimalPipe } from '@angular/common';
 import { ExternalLinkDirective } from '../../directives/external-link.directive';
 import { locationArray, PurposeOfSite, purposeOfSiteArray, WorldLocation } from '../../types/carbon-estimator';
 import { CarbonIntensityService } from '../../services/carbon-intensity.service';
-import { DownstreamEmissionsEstimator } from '../../estimation/estimate-downstream-emissions';
-import { CLOUD_AVERAGE_PUE, ON_PREMISE_AVERAGE_PUE } from '../../estimation/constants';
-import { desktop, laptop, mobile, monitor, network, server, tablet } from '../../estimation/device-type';
+import { DownstreamEmissionsEstimator } from '../../services/estimation/estimate-downstream-emissions';
+import { CLOUD_AVERAGE_PUE, ON_PREMISE_AVERAGE_PUE } from '../../services/estimation/constants';
+import { desktop, laptop, mobile, monitor, network, server, tablet } from '../../services/estimation/device-type';
 
 const purposeDescriptions: Record<PurposeOfSite, string> = {
   information: 'Information',
