+++ src/app/services/estimation/device-type.spec.ts
