+++ src/app/services/estimation/estimate-direct-emissions.spec.ts
@@ -1,5 +1,5 @@
-import { DeviceCategory } from '../types/carbon-estimator';
-import { KgCo2e } from '../types/units';
+import { DeviceCategory } from '../../types/carbon-estimator';
+import { KgCo2e } from '../../types/units';
 import { DeviceUsage } from './device-usage';
 import { estimateDirectEmissions } from './estimate-direct-emissions';
