+++ src/app/services/estimation/estimate-downstream-emissions.spec.ts
@@ -1,6 +1,6 @@
-import { provideFakeCO2CalculatorMock } from '../test-utils/fake-co2-calculator.testing';
-import { Downstream, DownstreamEstimation, PurposeOfSite, basePurposeArray } from '../types/carbon-estimator';
-import { sumValues } from '../utils/number-object';
+import { provideFakeCO2CalculatorMock } from '../../test-utils/fake-co2-calculator.testing';
+import { Downstream, DownstreamEstimation, PurposeOfSite, basePurposeArray } from '../../types/carbon-estimator';
+import { sumValues } from '../../utils/number-object';
 import { DownstreamEmissionsEstimator } from './estimate-downstream-emissions';
 import { TestBed } from '@angular/core/testing';
