+++ src/app/services/estimation/estimate-upstream-emissions.spec.ts
@@ -1,5 +1,5 @@
-import { DeviceCategory } from '../types/carbon-estimator';
-import { KgCo2e } from '../types/units';
+import { DeviceCategory } from '../../types/carbon-estimator';
+import { KgCo2e } from '../../types/units';
 import { DeviceUsage } from './device-usage';
 import { estimateUpstreamEmissions } from './estimate-upstream-emissions';
